+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   Support additional `<RMap>` properties, including disabling of the rotation
 -   Fix [#109], interactions do not support JSX styles
+-   Add event callbacks to `RDraw` and `RModify`
 
 ### [1.4.7] 2022-12-23